backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
SQLAlchemy==2.0.35
psycopg[binary]==3.2.3
alembic==1.13.2
passlib[bcrypt]==1.7.4
pydantic==2.9.2
pydantic-settings==2.5.2
PyJWT==2.9.0
python-jose[cryptography]
passlib[bcrypt]
